requirements.txt
streamlit 
phidata
openai
google-search-results
SpeechRecognition
pyaudio